feat: shop manage

+++ coleslaw/agency_manage/views/shop_manage_views/shop_manage_views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse
 from django.views.generic import View
 from django.http import HttpRequest, JsonResponse
@@ -9,20 +9,29 @@ from django.utils import timezone
 from django.db import transaction
 from system_manage.decorators import permission_required
 from system_manage.models import Shop, ShopCategory, ShopTable, Agency, AgencyShop
+from agency_manage.views.agency_manage_views.auth_views import check_agency
+
 
 class ShopManageView(View):
     '''
         가맹점 관리 화면
     '''
-    @method_decorator(permission_required(redirect_url='system_manage:denied'))
+    @method_decorator(permission_required(redirect_url='agency_manage:denied'))
     def get(self, request: HttpRequest, *args, **kwargs):
         context = {}
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return redirect('agency_manage:notfound')
+        context['agency'] = agency
+        
         paginate_by = '20'
         page = request.GET.get('page', '1')
         search_type = request.GET.get('search_type', '')
         search_keyword = request.GET.get('search_keyword', '')
 
         filter_dict = {}
+        filter_dict['agency'] = agency
 
         if search_keyword:
             context['search_type'] = search_type
@@ -55,26 +64,33 @@ class ShopManageView(View):
         context['pagelist'] = pagelist
         context['page_obj'] = page_obj
 
-        return render(request, 'shop_manage/shop_manage.html', context)
+        return render(request, 'agency_shop_manage/shop_manage.html', context)
     
 
 class ShopCreateView(View):
     '''
         가맹점 생성
     '''
-    @method_decorator(permission_required(redirect_url='system_manage:denied'))
+    @method_decorator(permission_required(redirect_url='agency_manage:denied'))
     def get(self, request: HttpRequest, *args, **kwargs):
         context = {}
-        shop_category = ShopCategory.objects.all().values('id', 'name').order_by('id')
-        agency = Agency.objects.all().values('id', 'name').order_by('id')
-        context['shop_category'] = shop_category
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return redirect('agency_manage:notfound')
         context['agency'] = agency
+        shop_category = ShopCategory.objects.all().values('id', 'name').order_by('id')
+        context['shop_category'] = shop_category
 
-        return render(request, 'shop_manage/shop_create.html', context)
+        return render(request, 'agency_shop_manage/shop_create.html', context)
     
     @method_decorator(permission_required(raise_exception=True))
     def post(self, request: HttpRequest, *args, **kwargs):
-        agency_id = request.POST['agency_id']
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return JsonResponse({'message': '데이터오류'}, status = 400)
+        
         shop_category_id = request.POST['shop_category_id']
 
         shop_name = request.POST['shop_name'].strip()
@@ -135,61 +151,77 @@ class ShopCreateView(View):
             return JsonResponse({'message': '등록 실패.'}, status=400)
 
 
-        return JsonResponse({'message' : '등록 되었습니다.', 'url':reverse("system_manage:shop_detail", kwargs={"pk" : shop.id})},  status = 202)
+        return JsonResponse({'message' : '등록 되었습니다.', 'url':reverse("agency_manage:shop_detail", kwargs={"agency_id":agency_id, "pk" : shop.id})},  status = 202)
     
 
 class ShopDetailView(View):
     '''
         가맹점 상세 화면
     '''
-    @method_decorator(permission_required(redirect_url='system_manage:denied'))
+    @method_decorator(permission_required(redirect_url='agency_manage:denied'))
     def get(self, request: HttpRequest, *args, **kwargs):
         context = {}
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return redirect('agency_manage:notfound')
+        context['agency'] = agency
         pk = kwargs.get("pk")
-        data = get_object_or_404(Shop, pk=pk)
+        data = get_object_or_404(Shop, pk=pk, agency=agency)
         context['data'] = data
 
-        return render(request, 'shop_manage/shop_detail.html', context)
+        return render(request, 'agency_shop_manage/shop_detail.html', context)
     
     @method_decorator(permission_required(raise_exception=True))
     def delete(self, request: HttpRequest, *args, **kwargs):
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return JsonResponse({'message': '데이터오류'}, status = 400)
         pk = kwargs.get('pk')
         try:
-            shop = Shop.objects.get(pk=pk)
+            shop = Shop.objects.get(pk=pk, agency=agency)
         except:
             return JsonResponse({"message": "데이터 오류"},status=400)
         
         shop.delete()
 
-        return JsonResponse({'message' : '삭제되었습니다.', 'url':reverse('system_manage:shop_manage')},  status = 202)
+        return JsonResponse({'message' : '삭제되었습니다.', 'url':reverse('agency_manage:shop_manage', kwargs={'agency_id':agency_id})},  status = 202)
     
 
 class ShopEditView(View):
     '''
         가맹점 수정 화면
     '''
-    @method_decorator(permission_required(redirect_url='system_manage:denied'))
+    @method_decorator(permission_required(redirect_url='agency_manage:denied'))
     def get(self, request: HttpRequest, *args, **kwargs):
         context = {}
-        pk = kwargs.get("pk")
-        data = get_object_or_404(Shop, pk=pk)
-        context['data'] = data
-        shop_category = ShopCategory.objects.all().values('id', 'name').order_by('id')
-        agency = Agency.objects.all().values('id', 'name').order_by('id')
-        context['shop_category'] = shop_category
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return redirect('agency_manage:notfound')
         context['agency'] = agency
 
-        return render(request, 'shop_manage/shop_edit.html', context)
+        pk = kwargs.get("pk")
+        data = get_object_or_404(Shop, pk=pk, agency=agency)
+        context['data'] = data
+        shop_category = ShopCategory.objects.all().values('id', 'name').order_by('id')
+        context['shop_category'] = shop_category
+
+        return render(request, 'agency_shop_manage/shop_edit.html', context)
     
     @method_decorator(permission_required(raise_exception=True))
     def post(self, request: HttpRequest, *args, **kwargs):
+        agency_id = kwargs.get('agency_id')
+        agency = check_agency(pk=agency_id)
+        if not agency:
+            return JsonResponse({'message': '데이터오류'}, status = 400)
         pk = kwargs.get('pk')
         try:
-            shop = Shop.objects.get(pk=pk)
+            shop = Shop.objects.get(pk=pk, agency=agency)
         except:
             return JsonResponse({"message": "데이터 오류"},status=400)
         
-        agency_id = request.POST['agency_id']
         shop_category_id = request.POST['shop_category_id']
         shop_name = request.POST['shop_name'].strip()
         description = request.POST['description'].strip()
@@ -205,11 +237,6 @@ class ShopEditView(View):
         entry_image = request.FILES.get("entry_image")
 
         try:
-            agency = Agency.objects.get(pk=agency_id)
-        except:
-            return JsonResponse({'message': '에이전시 오류 입니다.'}, status=400)
-
-        try:
             shop_category = ShopCategory.objects.get(pk=shop_category_id)
         except:
             return JsonResponse({'message': '카테고리 오류 입니다.'}, status=400)
@@ -218,7 +245,6 @@ class ShopEditView(View):
         if Shop.objects.filter(name=shop_name).exclude(pk=shop.pk).exists():
             return JsonResponse({'message': '이미 존재하는 가맹점 명 입니다.'}, status=400)
         
-        shop.agency = agency
         shop.shop_category = shop_category
         shop.name = shop_name
         shop.description = description
@@ -236,6 +262,6 @@ class ShopEditView(View):
             shop.entry_image = entry_image
         shop.save()
 
-        return JsonResponse({'message' : '수정 되었습니다.', 'url':reverse("system_manage:shop_detail", kwargs={"pk" : shop.id})},  status = 202)
+        return JsonResponse({'message' : '수정 되었습니다.', 'url':reverse("agency_manage:shop_detail", kwargs={"agency_id":agency_id, "pk" : shop.id})},  status = 202)
